Support targeting alternative element for style injection
The goober package implicitly appends styles to document.head
but offers a way to rebind `css` and `styled` to another element.

This prevents co-locating styles for shadow dom applications.

This change adds a layer of indirection to goober and a utility method to globally
target another element for styling injection.